Make GeoPointBridge et al. public
... but with private constructors: the builders are still needed to
instantiate the bridge and its markers.

This is necessary so that the bridge can implement
BridgeBuilder<TypeBridge & PropertyBridge, ...> in the following commit.
