--- mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/bridge/builtin/spatial/impl/GeoPointBridgeImpl.java
@@ -1,100 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.v6poc.entity.pojo.bridge.builtin.spatial.impl;
-
-import java.util.Objects;
-import java.util.function.Function;
-import java.util.stream.Collector;
-
-import org.hibernate.search.v6poc.backend.document.DocumentElement;
-import org.hibernate.search.v6poc.backend.document.model.Store;
-import org.hibernate.search.v6poc.backend.document.model.IndexSchemaElement;
-import org.hibernate.search.v6poc.backend.document.IndexFieldAccessor;
-import org.hibernate.search.v6poc.backend.spatial.GeoPoint;
-import org.hibernate.search.v6poc.backend.spatial.ImmutableGeoPoint;
-import org.hibernate.search.v6poc.entity.model.SearchModel;
-import org.hibernate.search.v6poc.entity.pojo.bridge.Bridge;
-import org.hibernate.search.v6poc.entity.pojo.model.PojoModelElement;
-import org.hibernate.search.v6poc.entity.pojo.model.PojoModelElementAccessor;
-import org.hibernate.search.v6poc.entity.pojo.model.PojoElement;
-import org.hibernate.search.v6poc.util.SearchException;
-import org.hibernate.search.v6poc.util.StreamHelper;
-
-/**
- * @author Yoann Rodiere
- */
-public class GeoPointBridgeImpl implements Bridge {
-
-	private final String fieldName;
-	private final Store store;
-	private final String markerSet;
-
-	private IndexFieldAccessor<GeoPoint> fieldAccessor;
-	private Function<PojoElement, GeoPoint> coordinatesExtractor;
-
-	public GeoPointBridgeImpl(String fieldName, Store store, String markerSet) {
-		this.fieldName = fieldName;
-		this.store = store;
-		this.markerSet = markerSet;
-	}
-
-	@Override
-	public void bind(IndexSchemaElement indexSchemaElement, PojoModelElement bridgedPojoModelElement,
-			SearchModel searchModel) {
-		if ( fieldName == null || fieldName.isEmpty() ) {
-			// TODO retrieve the default name somehow when parameters.name() is empty
-			throw new UnsupportedOperationException( "Default field name not implemented yet" );
-		}
-
-		fieldAccessor = indexSchemaElement.field( fieldName ).asGeoPoint().store( store ).createAccessor();
-
-		if ( bridgedPojoModelElement.isAssignableTo( GeoPoint.class ) ) {
-			PojoModelElementAccessor<GeoPoint> sourceAccessor = bridgedPojoModelElement.createAccessor( GeoPoint.class );
-			coordinatesExtractor = sourceAccessor::read;
-		}
-		else {
-			PojoModelElementAccessor<Double> latitudeAccessor = bridgedPojoModelElement.properties()
-					.filter( model -> model.markers( LatitudeMarker.class )
-							.anyMatch( m -> Objects.equals( markerSet, m.getMarkerSet() ) ) )
-					.collect( singleMarkedProperty( "latitude", fieldName, markerSet ) )
-					.createAccessor( Double.class );
-			PojoModelElementAccessor<Double> longitudeAccessor = bridgedPojoModelElement.properties()
-					.filter( model -> model.markers( LongitudeMarker.class )
-							.anyMatch( m -> Objects.equals( markerSet, m.getMarkerSet() ) ) )
-					.collect( singleMarkedProperty( "longitude", fieldName, markerSet ) )
-					.createAccessor( Double.class );
-
-			coordinatesExtractor = bridgedElement -> {
-				Double latitude = latitudeAccessor.read( bridgedElement );
-				Double longitude = longitudeAccessor.read( bridgedElement );
-
-				if ( latitude == null || longitude == null ) {
-					return null;
-				}
-
-				return new ImmutableGeoPoint( latitude, longitude );
-			};
-		}
-	}
-
-	private static Collector<PojoModelElement, ?, PojoModelElement> singleMarkedProperty(
-			String markerName, String fieldName, String markerSet) {
-		return StreamHelper.singleElement(
-				() -> new SearchException( "Could not find a property with the " + markerName
-						+ " marker for field '" + fieldName + "' (marker set: '" + markerSet + "')" ),
-				() -> new SearchException( "Found multiple properties with the " + markerName
-						+ " marker for field '" + fieldName + "' (marker set: '" + markerSet + "')" )
-				);
-	}
-
-	@Override
-	public void write(DocumentElement target, PojoElement source) {
-		GeoPoint coordinates = coordinatesExtractor.apply( source );
-		fieldAccessor.write( target, coordinates );
-	}
-
-}
