--- mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/bridge/builtin/spatial/impl/LongitudeMarker.java
@@ -1,20 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.v6poc.entity.pojo.bridge.builtin.spatial.impl;
-
-public class LongitudeMarker {
-
-	private final String markerSet;
-
-	public LongitudeMarker(String markerSet) {
-		this.markerSet = markerSet;
-	}
-
-	public String getMarkerSet() {
-		return markerSet;
-	}
-}
